pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "skill-capture"
version = "1.0.0"
description = "A privacy-first AI agent that learns your workflows and turns them into one-click Skills."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "SkillCapture Contributors" }
]
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "python-frontmatter>=1.1.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0"
]

[project.scripts]
skill-capture-cli = "skill_capture.cli:main"
skill-capture-mcp = "skill_capture.server:main"

[tool.setuptools]
packages = ["skill_capture", "skill_capture.core"]
